BenchmarkOperation abstraction
